Add form with table. In order to be able to use thymeleaf...
I had to understand that when declaring the th:object="${tripDetailsForm}" I was enabling it so that
on the th:field properties I was able to call the inner part of the tripDetailsForm by
using the star symbol. So, in the expression th:field=="*{tripDetails[__${rowStat.index}__].tripId}"
I'm accessing the element in a way such as tripDetailsForm.tripDetail[index].tripId. The star sign
serves as a long distance connector from the th:object to the variable I'm about to handle.
In this case, it was a list, but it could be a simpler object.